+++ core/netft/src/client.cpp
@@ -17,7 +17,7 @@ public:
 
   ~DeferredDestroyer() {
     {
-      std::lock_guard<std::mutex> lock(mutex_);
+      std::scoped_lock lock(mutex_);
       stopping_ = true;
     }
     condition_.notify_one();
@@ -26,7 +26,7 @@ public:
 
   void enqueue(std::unique_ptr<Value> value) {
     {
-      std::lock_guard<std::mutex> lock(mutex_);
+      std::scoped_lock lock(mutex_);
       pending_.push_back(std::move(value));
     }
     condition_.notify_one();
@@ -58,8 +58,7 @@ private:
 
 } // namespace
 
-Client::Client(Config config)
-    : impl_(std::make_unique<Impl>(std::move(config))) {}
+Client::Client(Config config) : impl_(std::make_unique<Impl>(std::move(config))) {}
 
 Client::~Client() {
   if (impl_ && impl_->called_from_worker_thread()) {
@@ -71,16 +70,13 @@ Client::~Client() {
   stop();
 }
 
-void Client::start(SampleCallback callback) {
-  impl_->start(std::move(callback));
-}
+void Client::start(SampleCallback callback) { impl_->start(std::move(callback)); }
 
 void Client::stop() noexcept { impl_->stop(); }
 
 void Client::bias() { impl_->bias(); }
 
-bool Client::wait_for_first_sample(
-    const std::chrono::duration<double> timeout) {
+bool Client::wait_for_first_sample(const std::chrono::duration<double> timeout) {
   return impl_->wait_for_first_sample(timeout);
 }
 
@@ -90,8 +86,6 @@ FaultCode Client::fault_code() const noexcept { return impl_->fault_code(); }
 
 HealthSnapshot Client::health() const { return impl_->health(); }
 
-std::optional<Sample> Client::latest_sample() const {
-  return impl_->latest_sample();
-}
+std::optional<Sample> Client::latest_sample() const { return impl_->latest_sample(); }
 
 } // namespace netft